requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.0.3
python-multipart==0.0.6
aiofiles==23.2.1
pymongo==4.6.0
motor==3.3.2
requests==2.31.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
httpx==0.25.2
PyJWT==2.8.0
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.3
typing-extensions==4.8.0